+++ class-list-generator/build.gradle.kts
@@ -8,5 +8,5 @@ repositories {
 
 dependencies {
     implementation(libs.classgraph)
-    implementation("dev.freya02:reflection-metadata-commons")
+    implementation("dev.freya02:classpath-scanner-commons")
 }
