+++ classpath-scanner-commons/build.gradle.kts
