+++ test-bot/src/test/resources/META-INF/BotCommands-restarter.properties
@@ -8,7 +8,7 @@ restart.exclude.restarter-prod-res=BotCommands-restarter/build/resources/main
 
 restart.exclude.core-prod=BotCommands-core/build/classes/(?:kotlin|java)/main
 restart.exclude.core-prod-res=BotCommands-core/build/resources/main
-restart.exclude.core-prod-generated-bins=BotCommands-core/build/generated/bins/reflection-metadata-commons
+restart.exclude.core-prod-generated-bins=BotCommands-core/build/generated/bins/classpath-scanner-commons
 
 restart.exclude.rate-limit-prod=BotCommands-rate-limit/build/classes/(?:kotlin|java)/main
 restart.exclude.rate-limit-prod-res=BotCommands-rate-limit/build/resources/main
